# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for Jingliange.",
        "title": "Jingliange Server API",
        "termsOfService": "TO IMPLEMENT",
        "contact": {},
        "version": "1.0"
    },
    "host": "49.234.22.169:8000",
    "paths": {
...
        "/api/v1/menu/comment": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户对菜品进行评论（需要JWT认证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "菜品评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "评论参数",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":\"comment added successfully\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "{\"code\":401,\"msg\":\"unauthorized\",\"data\":\"Token required\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "404": {
                        "description": "{\"code\":404,\"msg\":\"not found\",\"data\":\"Menu item not found\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/getComments": {
            "post": {
                "description": "获取指定菜品的评论列表，支持分页",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取菜品评论列表",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuCommentsQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"id\":1,\"menu_id\":1,\"user_id\":\"user123\",\"preference\":1,\"comment\":\"非常好吃！\",\"status\":1}]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.MenuFeedback"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "404": {
                        "description": "{\"code\":404,\"msg\":\"not found\",\"data\":\"Menu item not found\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/getLikeStatus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取用户对特定菜品的点赞状态（需要JWT认证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取菜品点赞状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询参数",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuLikeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"liked\":true,\"preference\":1}}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "{\"code\":401,\"msg\":\"unauthorized\",\"data\":\"Token required\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "404": {
                        "description": "{\"code\":404,\"msg\":\"not found\",\"data\":\"Menu item not found\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/getMenu": {
            "post": {
                "description": "获取全部净莲阁菜单，返回菜单项列表，每个菜单项包含名称、图片 url、营养价值表 json、点赞数等信息。输入名称过滤菜单项，支持模糊匹配。如果提供了JWT token，会返回用户的点赞状态。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取净莲阁的菜单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token (可选)",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "查询参数",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"id\":1,\"name\":\"紫菜汤\",\"image_url\":\"/images/menu.jpg\",\"desc\":\"美味的菜品\",\"nutrition\":\"...\",\"ingredients\":\"...\",\"status\":1,\"like_count\":5,\"liked\":true}]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.MenuWithUserLikes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/getMenuByID": {
            "post": {
                "description": "根据菜品ID获取单个菜品的详细信息，包括点赞数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取单个菜品信息",
                "parameters": [
                    {
                        "description": "查询参数",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuByIDRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"id\":1,\"name\":\"紫菜汤\",\"image_url\":\"/images/menu.jpg\",\"desc\":\"美味的菜品\",\"nutrition\":\"...\",\"ingredients\":\"...\",\"status\":1,\"like_count\":5,\"create_time\":\"2012-1-1\",\"update_time\":\"2012-1-1\"}}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.MenuWithLikes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "404": {
                        "description": "{\"code\":404,\"msg\":\"not found\",\"data\":\"Menu item not found\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/menu/like": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "用户为喜欢的菜品点赞或取消点赞（需要JWT认证）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "菜品点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "点赞参数",
                        "name": "like",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MenuLikeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":\"liked successfully\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "400": {
                        "description": "{\"code\":400,\"msg\":\"invalid params\",\"data\":\"Invalid input data\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "401": {
                        "description": "{\"code\":401,\"msg\":\"unauthorized\",\"data\":\"Token required\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "404": {
                        "description": "{\"code\":404,\"msg\":\"not found\",\"data\":\"Menu item not found\"}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    },
                    "500": {
                        "description": "{\"code\":500,\"msg\":\"internal server error\",\"data\":null}",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "model.MenuByIDRequest": {
            "type": "object",
            "properties": {
                "menu_id": {
                    "description": "菜品ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.MenuCommentRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "评论内容",
                    "type": "string",
                    "example": "非常好吃的菜品！"
                },
                "menu_id": {
                    "description": "菜品ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.MenuCommentsQueryRequest": {
            "type": "object",
            "properties": {
                "menu_id": {
                    "description": "菜品ID",
                    "type": "integer",
                    "example": 1
                },
                "page_number": {
                    "description": "页码，从零开始",
                    "type": "integer",
                    "example": 0
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "model.MenuFeedback": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "非常好吃"
                },
                "create_time": {
                    "description": "Creation time",
                    "type": "string",
                    "example": "2012-1-1"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "menu_id": {
                    "description": "菜单 ID",
                    "type": "integer",
                    "example": 1
                },
                "preference": {
                    "description": "状态：0 默认，1 喜欢，2 不喜欢",
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "description": "状态：0 删除，1 正常",
                    "type": "integer",
                    "example": 1
                },
                "update_time": {
                    "description": "Update time",
                    "type": "string",
                    "example": "2012-1-1"
                },
                "user_id": {
                    "description": "用户\t ID",
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "model.MenuLikeRequest": {
            "type": "object",
            "properties": {
                "menu_id": {
                    "description": "菜品ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "model.MenuQueryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "菜单名称，支持模糊匹配",
                    "type": "string",
                    "example": "紫菜汤"
                },
                "page_number": {
                    "description": "页码，从零开始",
                    "type": "integer",
                    "example": 0
                },
                "page_size": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 10
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}